Major Refactoring
This update includes some major refactoring of the code to allow for
overall better readability and maintenance.
Includes:
*Dialogs have been changed to now have three classes. The base class has
only an close button and a window with a layout. The title class
inherits the base class and has titles. The content class inherits the
title class and has content. This allows for much cleaner code and all
that is needed to make a dialog is to feed one of the classes a couple
of parameters.
*Better and more consistent documentation overall.
*Many functions were split into smaller units to allow for better
readability.
Todo:
*MainActivity java file is fairly large and consists of many sections.
It may be beneficial to split each section into a seperate files.
*There are still a few areas that could possibly be split into a
seperate file such as preferences, and spinners.